Fix imports, exports, and references to imported things so that this will finally compile! ...except not quite, turns out there's a function called mapping() that I missed and will have to deal with.  Also, I moved the definition of StorageLength (and the corresponding utility functions) StorageLength from truffle-decode-utils to truffle-decoder.  So, still don't use this, it still won't compile.
